Move utils package inside core for consolidated source code
