+++ modules/lisk-core-p2p/package.json
@@ -9,7 +9,6 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@lisk/core-utils": "^1.0.0",
     "@lisk/lisk-logger-bunyan": "^1.0.0"
   }
 }
