--- packages/lisk-core-utils/.babelrc
@@ -1,9 +1,0 @@
-{
-	"presets": ["env"],
-	"plugins": ["transform-runtime"],
-	"env": {
-		"test": {
-			"plugins": ["istanbul"]
-		}
-	}
-}
