--- packages/lisk-core-utils/.editorconfig
@@ -1,8 +1,0 @@
-root = true
-
-[*]
-charset = utf-8
-end_of_line = lf
-indent_style = tab
-insert_final_newline = true
-trim_trailing_whitespace = true
