--- packages/lisk-core-utils/.eslintignore
@@ -1,2 +1,0 @@
-coverage/*
-dist/*
