--- packages/lisk-core-utils/.eslintrc.json
@@ -1,4 +1,0 @@
-{
-	"extends": ["lisk-base"],
-	"plugins": ["mocha"]
-}
