--- packages/lisk-core-utils/.gitignore
@@ -1,44 +1,0 @@
-# General
-.DS_Store
-*.swp
-*.orig
-
-# Logs
-logs
-*.log
-npm-debug.log*
-
-# Runtime data
-pids
-*.pid
-*.seed
-
-# Coverage directory used by tools like istanbul
-coverage
-
-# nyc test coverage
-.nyc_output
-
-# node-waf configuration
-.lock-wscript
-
-# Compiled binary addons (http://nodejs.org/api/addons.html)
-build/Release
-
-# Dependency directories
-node_modules
-jspm_packages
-
-# Optional npm cache directory
-.npm
-
-# Optional REPL history
-.node_repl_history
-
-# IDE directories
-.vscode
-.idea
-tmp
-
-# Distribution directory
-dist
