--- packages/lisk-core-utils/.lintstagedrc.json
@@ -1,4 +1,0 @@
-{
-	"*.js": ["eslint", "prettier --write", "git add"],
-	"*.{json,md}": ["prettier --write", "git add"]
-}
