--- packages/lisk-core-utils/.node-version
@@ -1,1 +1,0 @@
-8.11.3
