--- packages/lisk-core-utils/.npmignore
@@ -1,19 +1,0 @@
-.babelrc
-.editorconfig
-.eslintignore
-.eslintrc.json
-.gitignore
-.npmignore
-.nvmrc
-.nycrc
-.snyk
-Jenkinsfile*
-package-lock.json
-yarn.lock
-
-.vscode/
-.idea/
-.nyc_output/
-coverage/
-src/
-test/
