--- packages/lisk-core-utils/.npmrc
@@ -1,2 +1,0 @@
-message = ":arrow_up: Version %s"
-save-exact = true
