--- packages/lisk-core-utils/.nycrc
@@ -1,5 +1,0 @@
-{
-	"require": ["babel-register"],
-	"sourceMap": false,
-	"instrument": false
-}
