--- packages/lisk-core-utils/.prettierignore
@@ -1,9 +1,0 @@
-package.json
-
-.nyc_output/
-coverage/
-dist/
-node_modules/
-
-# BUG: See https://github.com/prettier/prettier/issues/3223
-docs/CONTRIBUTING.md
