--- packages/lisk-core-utils/.prettierrc.json
@@ -1,5 +1,0 @@
-{
-	"singleQuote": true,
-	"trailingComma": "all",
-	"useTabs": true
-}
