--- packages/lisk-core-utils/Jenkinsfile.nightly
@@ -1,23 +1,0 @@
-pipeline {
-	agent { node { label 'lisk-core-utils2' } }
-	stages {
-		stage('Prepare workspace') {
-			steps {
-				checkout([$class: 'GitSCM', branches: [[name: '*/master']], doGenerateSubmoduleConfigurations: false, extensions: [], submoduleCfg: [], userRemoteConfigs: [[credentialsId: 'scan-1', url: 'https://github.com/LiskHQ/lisk-core-utils2']]])
-			}
-		}
-		stage('Install dependencies') {
-			steps {
-				sh 'npm install --verbose'
-			}
-		}
-		stage('Run vulnerabilities check') {
-			steps {
-				withCredentials([string(credentialsId: 'liskhq-snyk-token', variable: 'SNYK_TOKEN')]) {
-					sh 'snyk test'
-				}
-				sh 'sha1sum package.json >~/.lisk-core-utils2.package.json.sha1sum'
-			}
-		}
-	}
-}
