--- packages/lisk-core-utils/Jenkinsfile.private
@@ -1,30 +1,0 @@
-pipeline {
-	agent { node { label 'lisk-core-utils2' } }
-	stages {
-		stage('Prepare workspace') {
-			steps {
-				checkout scm
-			}
-		}
-		stage('Install dependencies') {
-			steps {
-				sh 'npm install --verbose'
-			}
-		}
-		stage('Run vulnerabilities check') {
-			steps {
-				withCredentials([string(credentialsId: 'liskhq-snyk-token', variable: 'SNYK_TOKEN')]) {
-					sh '''
-					if [ -r ~/.lisk-core-utils2.package.json.sha1sum ]; then
-						if sha1sum --status --check ~/.lisk-core-utils2.package.json.sha1sum; then
-							echo "Skipping vulnerabilities check since package.json has not changed."
-							exit 0
-						fi
-					fi
-					snyk test
-					'''
-				}
-			}
-		}
-	}
-}
