--- packages/lisk-core-utils/docs/ISSUE_TEMPLATE.md
@@ -1,7 +1,0 @@
-### Expected behavior
-
-### Actual behavior
-
-### Steps to reproduce
-
-### Which version(s) does this affect? (Environment, OS, etc...)
