--- packages/lisk-core-utils/docs/PULL_REQUEST_TEMPLATE.md
@@ -1,16 +1,0 @@
-### What was the problem?
-
-### How did I fix it?
-
-### How to test it?
-
-### Review checklist
-
-* The PR resolves #INSERT_ISSUE_NUMBER
-* All new code is covered with unit tests
-* All new code was formatted with Prettier
-* Linting passes
-* Tests pass
-* Commit messages follow the
-	[commit guidelines](CONTRIBUTING.md#git-commit-messages)
-* Documentation has been added/updated
