--- packages/lisk-core-utils/src/index.js
@@ -1,13 +1,0 @@
-const ChannelFactory = require('./factories/channels');
-const ModuleFactory = require('./factories/modules');
-const ComponentFactory = require('./factories/components');
-const Event = require('./event');
-const Action = require('./action');
-
-module.exports = {
-	ChannelFactory,
-	ModuleFactory,
-	ComponentFactory,
-	Event,
-	Action,
-};
