--- packages/lisk-core-utils/test/.eslintrc.json
@@ -1,17 +1,0 @@
-{
-	"env": {
-		"mocha": true
-	},
-	"globals": {
-		"Given": true,
-		"When": true,
-		"Then": true,
-		"expect": true,
-		"sandbox": true,
-		"sinon": true
-	},
-	"rules": {
-		"arrow-body-style": "off",
-		"new-cap": ["error", { "capIsNewExceptions": ["Given", "When", "Then"] }]
-	}
-}
