--- packages/lisk-core-utils/test/_global_hooks.js
@@ -1,19 +1,0 @@
-/*
- * LiskHQ/lisk-core-utils2
- * Copyright © 2017 Lisk Foundation
- *
- * See the LICENSE file at the top-level directory of this distribution
- * for licensing information.
- *
- * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
- * no part of this software, including this file, may be copied, modified,
- * propagated, or distributed except according to the terms contained in the
- * LICENSE file.
- *
- * Removal or modification of this copyright notice is prohibited.
- *
- */
-/* eslint-disable mocha/no-top-level-hooks */
-afterEach(() => {
-	return sandbox.restore();
-});
