--- packages/lisk-core-utils/test/mocha.opts
@@ -1,5 +1,0 @@
---require babel-register
---require ./test/setup.js
---reporter spec
---recursive
---ui bdd
