--- packages/lisk-core-utils/test/steps/.eslintrc.json
@@ -1,5 +1,0 @@
-{
-	"rules": {
-		"import/prefer-default-export": "off"
-	}
-}
