--- packages/lisk-core-utils/test/steps/setup.js
@@ -1,24 +1,0 @@
-/*
- * LiskHQ/lisk-core-utils2
- * Copyright © 2017 Lisk Foundation
- *
- * See the LICENSE file at the top-level directory of this distribution
- * for licensing information.
- *
- * Unless otherwise agreed in a custom licensing agreement with the Lisk Foundation,
- * no part of this software, including this file, may be copied, modified,
- * propagated, or distributed except according to the terms contained in the
- * LICENSE file.
- *
- * Removal or modification of this copyright notice is prohibited.
- *
- */
-export function setUpIndex() {
-	// Stubs, spies, etc should be created here
-	// Environmental variables can be stored and/or updated
-}
-
-export function tearDownIndex() {
-	// Environmental variables can be restored here
-	// Sinon helpers which aren’t part of the sandbox should be restored here
-}
