+++ packages/lisk-core/package.json
@@ -17,10 +17,9 @@
     "url": "https://github.com/LiskHQ/lisk-core2/issues"
   },
   "engines": {
-	"node": ">=6.14.1 <=8.11.3",
-	"npm": ">=3.10.10 <=5.6.0"
+    "node": ">=6.14.1 <=8.11.3",
+    "npm": ">=3.10.10 <=5.6.0"
   },
-  "main": "src/index.js",
   "scripts": {
     "format": "prettier --write \"**/*.{js,json,md}\"",
     "lint": "eslint .",
@@ -37,18 +36,18 @@
     "postinstall": "npm run check:dependencies"
   },
   "bin": {
-	"liskctrl": "./bin/liskctrl"
+    "liskctrl": "./bin/liskctrl"
   },
   "dependencies": {
+    "@lisk/lisk-logger-bunyan": "^1.0.0",
+    "axon": "^2.0.3",
+    "axon-rpc": "^0.0.3",
     "babel-runtime": "6.26.0",
-	"@lisk/core-utils": "^1.0.0",
-	"@lisk/lisk-logger-bunyan": "^1.0.0",
-	"axon": "^2.0.3",
-	"axon-rpc": "^0.0.3",
-	"bluebird": "^3.5.1",
-	"commander": "^2.16.0",
-	"eventemitter2": "^5.0.1",
-	"fs-extra": "^7.0.0"
+    "bluebird": "^3.5.1",
+    "commander": "^2.16.0",
+    "debug": "^3.1.0",
+    "eventemitter2": "^5.0.1",
+    "fs-extra": "^7.0.0"
   },
   "devDependencies": {
     "babel-cli": "6.26.0",
