+++ packages/lisk-core/src/channels/base.js
