+++ packages/lisk-core/src/channels/event_emitter.js
@@ -19,11 +19,15 @@ module.exports = class EventEmitterChannel extends BaseChannel {
 	}
 
 	subscribe(eventName, cb) {
-		this.bus.on(new Event(eventName).key(), data => setImmediate(cb, Event.deserialize(data)));
+		this.bus.on(new Event(eventName).key(), data =>
+			setImmediate(cb, Event.deserialize(data)),
+		);
 	}
 
 	once(eventName, cb) {
-		this.bus.once(new Event(eventName).key(), data => setImmediate(cb, Event.deserialize(data)));
+		this.bus.once(new Event(eventName).key(), data =>
+			setImmediate(cb, Event.deserialize(data)),
+		);
 	}
 
 	publish(eventName, data) {
