+++ packages/lisk-core/src/event.js
