+++ packages/lisk-core/src/factories/channels.js
