+++ packages/lisk-core/src/factories/components.js
