+++ packages/lisk-core/src/factories/modules.js
