+++ packages/lisk-core/src/loaders/child_process.js
