+++ packages/lisk-core/src/modules/base.js
