+++ packages/lisk-core/src/modules/in_memory.js
